myGallery/requirements.txt
asgiref==3.8.1
Django==5.0.3
mysqlclient==2.1.1
Pillow==9.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
